more fixes for new faction
git-svn-id: svn://192.168.1.10/public/CoffeeMud@17907 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Common/interfaces/Faction.java
@@ -1467,7 +1467,7 @@ public interface Faction extends CMCommon, MsgListener, Contingent
 		 * @see com.planet_ink.coffee_mud.Abilities.interfaces.Ability#FLAG_DESCS
 		 * @return a bitmask of Ability flags that must not be set by the ability
 		 */
-		public int notflag();
+		public long notflag();
 
 		/**
 		 * A bitmask of ability flags that MUST be set for this usage to apply to an ability
@@ -1476,7 +1476,7 @@ public interface Faction extends CMCommon, MsgListener, Contingent
 		 * @see com.planet_ink.coffee_mud.Abilities.interfaces.Ability#FLAG_DESCS
 		 * @return a bitmask of Ability flags that must be set by the ability
 		 */
-		public int flag();
+		public long flag();
 
 		/**
 		 * Whether the abilityFlags() method is possibly a specific Ability ID
